feat(ci): #DEVOPS-4001 update cron schedule and install additional tools
The changes in this commit update the cron schedule for the Docker publish workflow and install additional tools required for the CI/CD pipeline.

The cron schedule has been updated from `26 1 * * *` to `15 1 * * *` to run the workflow at a different time.

The Dockerfile has been updated to install the following additional tools:
- curl, openssl, docker-cli, git, openssh-client, yq, ca-certificates
- Trivy v0.55.2
- Helm v3.16.3
- ArgoCD v2.14.7
- kubectl v1.31.2
- hadolint v2.12.0
- k6 v0.51.0
- GitHub CLI (gh) v2.49.0

The Dockerfile also creates an Azure DevOps agent user, sets up the Docker group, and configures sudo for the agent user. The working directory is set to `/azp`, and the necessary environment variables for Azure DevOps are set.

These changes ensure that the CI/CD pipeline has access to the required tools and dependencies to perform various tasks, such as container image scanning, Helm chart management, and Kubernetes deployment.